+++ endpoint-tapir/src/main/scala/busymachines/pureharm/endpoint/PureharmEndpointAllImplicits.scala
@@ -1,0 +1,3 @@
+package busymachines.pureharm.endpoint
+
+trait PureharmEndpointAllImplicits extends PureharmSproutTypeRestTapirImplicits