+++ endpoint-tapir/src/main/scala/busymachines/pureharm/endpoint/PureharmEndpointAllTypeAliases.scala
@@ -1,0 +1,5 @@
+package busymachines.pureharm.endpoint
+
+trait PureharmEndpointAllTypeAliases
+  extends sttp.tapir.Tapir with aliases.PureharmTapirAliases with aliases.PureharmTapirModelAliases
+  with aliases.PureharmTapirServerAliases {}